requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
plotly==5.17.0

# Web Interface
streamlit==1.28.2

# Configuration
python-dotenv==1.0.0

# AI/ML Framework - LangChain
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==1.0.1
langgraph==0.0.26

# Google Gemini Integration
google-generativeai==0.3.2

# Utilities
pydantic==2.5.3
tiktoken==0.5.2
